move createMuiTheme to App/index
