+++ src/Components/Login/index.js
@@ -2,20 +2,6 @@ import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 import './login.css';
 import Button from '@material-ui/core/Button';
-import { createMuiTheme, MuiThemeProvider } from '@material-ui/core';
-
-const theme = createMuiTheme({
-  // because we apparently use typography and it will be deprecated with the next release,
-  // to have a smooth transition:
-  typography: {
-    useNextVariants: true,
-  },
-  palette: {
-    primary: {
-      main: '#0a45c1',
-    },
-  },
-});
 
 @withRouter
 export default class Login extends Component {
@@ -75,11 +61,9 @@ export default class Login extends Component {
                       </td>
                       <td>
                         <br />
-                        <MuiThemeProvider theme={theme}>
-                          <Button variant="contained" color="primary" onClick={this.onSubmit}>
+                        <Button variant="contained" color="primary" onClick={this.onSubmit}>
                             Log in
-                          </Button>
-                        </MuiThemeProvider>
+                        </Button>
                       </td>
                     </tr>
                   </tbody>
