+++ src/Components/Navbar/index.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import './navbar.css';
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
@@ -32,22 +31,6 @@ export default class Navbar extends Component {
     );
   }
 }
-
-const theme = createMuiTheme({
-  // because we apparently use typography and it will be deprecated with the next release,
-  // to have a smooth transition:
-  typography: {
-    useNextVariants: true,
-  },
-  palette: {
-    primary: {
-      // light: will be calculated from palette.primary.main,
-      main: '#0a45c1',
-      // dark: will be calculated from palette.primary.main,
-      // contrastText: will be calculated to contrast with palette.primary.main
-    },
-  },
-});
 
 @QueryHOC(query)
 @withRouter
@@ -83,53 +66,51 @@ export class NavbarContent extends Component {
 
     return (
       <div className='navbar'>
-        <MuiThemeProvider theme={theme}>
-          <AppBar position="static">
-            <Toolbar>
-              <Typography variant="h6" color="inherit" className='typography'>
+        <AppBar position="static">
+          <Toolbar>
+            <Typography variant="h6" color="inherit" className='typography'>
                 Welkom, {user.firstName}
-              </Typography>
-              <div className='menuRight'>
-                <div>
-                  <IconButton
-                    className='button'
-                    color="inherit"
-                    aria-label="Menu"
-                  >
-                    <MenuIcon />
-                  </IconButton>
-                </div>
-                <div>
-                  <IconButton
-                    aria-owns={open ? 'menu-appbaraccount' : undefined}
-                    aria-haspopup="true"
-                    onClick={this.handleMenu}
-                    color="inherit"
-                  >
-                    <AccountCircle />
-                  </IconButton>
-                  <Menu
-                    id="menu-appbaraccount"
-                    anchorEl={anchorEl}
-                    anchorOrigin={{
-                      vertical: 'top',
-                      horizontal: 'right',
-                    }}
-                    transformOrigin={{
-                      vertical: 'top',
-                      horizontal: 'right',
-                    }}
-                    open={open}
-                    onClose={this.handleClose}
-                  >
-                    <MenuItem onClick={this.handleClose}>Account</MenuItem>
-                    <MenuItem onClick={this.logout}>Log uit</MenuItem>
-                  </Menu>
-                </div>
+            </Typography>
+            <div className='menuRight'>
+              <div>
+                <IconButton
+                  className='button'
+                  color="inherit"
+                  aria-label="Menu"
+                >
+                  <MenuIcon />
+                </IconButton>
               </div>
-            </Toolbar>
-          </AppBar>
-        </MuiThemeProvider>
+              <div>
+                <IconButton
+                  aria-owns={open ? 'menu-appbaraccount' : undefined}
+                  aria-haspopup="true"
+                  onClick={this.handleMenu}
+                  color="inherit"
+                >
+                  <AccountCircle />
+                </IconButton>
+                <Menu
+                  id="menu-appbaraccount"
+                  anchorEl={anchorEl}
+                  anchorOrigin={{
+                    vertical: 'top',
+                    horizontal: 'right',
+                  }}
+                  transformOrigin={{
+                    vertical: 'top',
+                    horizontal: 'right',
+                  }}
+                  open={open}
+                  onClose={this.handleClose}
+                >
+                  <MenuItem onClick={this.handleClose}>Account</MenuItem>
+                  <MenuItem onClick={this.logout}>Log uit</MenuItem>
+                </Menu>
+              </div>
+            </div>
+          </Toolbar>
+        </AppBar>
       </div>
     );
   }
