+++ src/Components/ProductRegistration/index.js
@@ -4,7 +4,6 @@ import gql from 'graphql-tag';
 import './productregistration.css';
 import Button from '@material-ui/core/Button';
 import AddIcon from '@material-ui/icons/Add';
-import { createMuiTheme, MuiThemeProvider } from '@material-ui/core';
 import Dialog from '@material-ui/core/Dialog/Dialog';
 import DialogTitle from '@material-ui/core/DialogTitle/DialogTitle';
 import DialogContent from '@material-ui/core/DialogContent/DialogContent';
@@ -18,19 +17,6 @@ const query = gql`{
     name code locations {code} recommended_stock 
   }
 }`;
-
-const theme = createMuiTheme({
-  // because we apparently use typography and it will be deprecated with the next release,
-  // to have a smooth transition:
-  typography: {
-    useNextVariants: true,
-  },
-  palette: {
-    primary: {
-      main: '#0a45c1',
-    },
-  },
-});
 
 const columnFormatting = ['name', 'code', ({ locations }) => locations.reduce((accum, { code }) => `${accum}, ${code}`, '').substring(2), 'recommended_stock'];
 const firstRowTable = ['Product', 'Type', 'Locatie', 'Aanbevolen voorraad'];
@@ -64,56 +50,54 @@ export default class ProductRegistration extends Component {
           <h3>
             Huidige producten:
           </h3>
-          <MuiThemeProvider theme={theme}>
-            <Button variant='fab' color='primary' className='add' onClick={this.handleClickOpen}>
-              <AddIcon />
-            </Button>
-            <Dialog
-              className='dialogueWindow'
-              open={this.state.open}
-              onClose={this.handleClose}
-              aria-labelledby='alert-dialog-title'
-              aria-describedby='alert-dialog-description'
-            >
-              <DialogTitle id='alert-dialog-title'>Voeg een product toe:</DialogTitle>
-              <DialogContent className='dialogueContent'>
-                <div>
-                  <TextField
-                    id='standard-name'
-                    label='Product'
-                    margin='normal'
-                  />
-                </div>
-                <div>
-                  <TextField
-                    id='standard-name'
-                    label='Type'
-                    margin='normal'
-                  />
-                </div>
+          <Button variant='fab' color='primary' className='add' onClick={this.handleClickOpen}>
+            <AddIcon />
+          </Button>
+          <Dialog
+            className='dialogueWindow'
+            open={this.state.open}
+            onClose={this.handleClose}
+            aria-labelledby='alert-dialog-title'
+            aria-describedby='alert-dialog-description'
+          >
+            <DialogTitle id='alert-dialog-title'>Voeg een product toe:</DialogTitle>
+            <DialogContent className='dialogueContent'>
+              <div>
                 <TextField
                   id='standard-name'
-                  label='Locatie'
+                  label='Product'
                   margin='normal'
                 />
-                <div>
-                  <TextField
-                    id='standard-name'
-                    label='Aanbevolen voorraad'
-                    margin='normal'
-                  />
-                </div>
-              </DialogContent>
-              <DialogActions>
-                <Button onClick={this.handleClose} color="primary">
+              </div>
+              <div>
+                <TextField
+                  id='standard-name'
+                  label='Type'
+                  margin='normal'
+                />
+              </div>
+              <TextField
+                id='standard-name'
+                label='Locatie'
+                margin='normal'
+              />
+              <div>
+                <TextField
+                  id='standard-name'
+                  label='Aanbevolen voorraad'
+                  margin='normal'
+                />
+              </div>
+            </DialogContent>
+            <DialogActions>
+              <Button onClick={this.handleClose} color="primary">
                   Cancel
-                </Button>
-                <Button onClick={this.handleClose} color="primary" autoFocus>
+              </Button>
+              <Button onClick={this.handleClose} color="primary" autoFocus>
                   Ok
-                </Button>
-              </DialogActions>
-            </Dialog>
-          </MuiThemeProvider>
+              </Button>
+            </DialogActions>
+          </Dialog>
         </div>
         <Table data = {data.items} headers = {firstRowTable} columns = {columnFormatting}/>
       </div>
